move firetv utils into utils package
